Split plat_win into legacy and modern
